==> notebooks/00_prise_en_main.ipynb <==
# Vérification du fonctionnement de Spark
1. Exécuter la cellule suivante

val rdd = sc.parallelize(0 to 999)
rdd.takeSample(false, 5)

# Téléchargement du jeu de données
1. Récupérer le jeu de données [Fichier des prénoms par départements de naissance (1900-2019)](https://www.data.gouv.fr/fr/datasets/ficher-des-prenoms-de-1900-a-2019/).
1. Faire de même avec la [Liste des départements](https://www.insee.fr/fr/information/4316069#titre-bloc-15) (format csv)
1. Extraire les fichiers des archives
1. Renommer le fichier des prénoms en `prenoms.txt` et celui des départements en `dpts.txt`.
