==> notebooks/05_preparation_des_donnees.ipynb <==
# Vérifier le contenu du fichier

val prenoms = sc.textFile("prenoms.txt")
prenoms.takeSample(false, 10)

# Extraire et sauvegarder un échantillon
1. Expliquer ce que fait la ligne suivante

prenoms.filter(l => l.startsWith("sexe") == false).filter(l => l.contains("XX") == false).sample(false, .001).saveAsTextFile("prenoms_sample.txt")

Cette requete permet de faire une lecture aleatoire tout en ignorant la premiere ligne (ligne contenant le mot sexe) ainsi que l'annee de naissance non renseignant ( valeur a xx). LE resultat de la lecture aleatoire est stocke dans le fichier prenoms_sample.txt, et ces donnees representent 1/1000 des donnees du fichier. 

# Vérifier l'échantillon

val echantillon = sc.textFile("prenoms_sample.txt")
println(echantillon.count)
println(echantillon.filter(l => l.startsWith("sexe")).count)
println(echantillon.filter(l => l.contains("XX")).count)
