==> notebooks/10_introduction_aux_rdd.ipynb <==
# Une fonction pour simplifier l'accès aux données

def extractField(s: String, fieldNumber: Int): String = {
    val fields = s.split(';')
    if (fieldNumber >= fields.length) "" else fields(fieldNumber)
}

def extractFieldS(s: String, fieldNumber: Int): String = {
    val fields = s.split(',')
    if (fieldNumber >= fields.length) "" else fields(fieldNumber)
}

println(extractField("2;CASSIOPEE;2009;33;3", 0))
println(extractField("2;CASSIOPEE;2009;33;3", 1))
println(extractField("2;CASSIOPEE;2009;33;3", 2))
println(extractField("2;CASSIOPEE;2009;33;3", 3))
println(extractField("2;CASSIOPEE;2009;33;3", 4))
println(extractField("2;CASSIOPEE;2009;33;3", 5))

# Charger les données
1. Créer le RDD `lignes` à partir du répertoire `prenoms_sample.txt`

val lignes = sc.textFile("prenoms_sample.txt")
lignes.take(10).foreach(println)

# Transformer les lignes en prénoms
1. En appliquant la méthode `map`, créer le RDD `prenoms` à partir de `lignes`

val prenoms = lignes.map(l => (
    extractField(l, 0).charAt(0),
    extractField(l, 1),
    extractField(l, 2).toInt,
    extractField(l, 3).toInt,
    extractField(l, 4).toDouble.toInt
))
prenoms.take(10).foreach(println)

# Interroger les données
La documentation des méthodes d'un RDD est disponible ([RDD](https://spark.apache.org/docs/latest/api/scala/index.html#org.apache.spark.rdd.RDD), [PairRDDFunctions](https://spark.apache.org/docs/latest/api/scala/index.html#org.apache.spark.rdd.PairRDDFunctions)).

1. Rappeler ce que sont les *transformations* et les *actions*
1. Donner, pour chaque prénom, son nombre d'occurences (`map` et `reduceByKey`)

//une transformation retourne un nouvel RDD par transformation du RDD courant
// une action déclenche le calcul d’une valeur sur un RDD
val prenom = sc.textFile("prenoms.txt")
val prenomMap = prenom.map(l=> (extractField(l, 1),1))
val prenomReduce = prenomMap.reduceByKey((l,x)=> (l+x))
prenomReduce.take(20).foreach(println)


1. Donner le nombre total de naissances avec un prénom féminin (`filter`, `map`, `reduce` ou `sum`)

//TODO
prenoms.filter(l => l._1=='2').map(l=>l._5).reduce((x,y) => x+y)

1. Donner l'effectif maximal et minimal par prénom (`map`, `aggregateByKey`)

//TODO
//val prenomEffect = prenoms.map(t=>(t._2, t._5)).aggregateByKey((Int.MaxValue,Int.MinValue))(
//    {case((m,n),c)=>(n min c, n max c)}, {case((m,n),(m1,n1))=>(m min m1, n max n1)})
                                 
//prenomEffect.take(10).foreach(println)
val result = lignes.map(l => (extractField(l, 1),extractField(l, 4).toDouble.toInt))
            .aggregateByKey((Int.MaxValue,Int.MinValue))(
{case ((k,u),v) => (k min v, u max v)},{case ((k,u),(k1, u1)) => (k min k1, u max u1)},                
            )
result.take(10).foreach(println)

1. Sur le modèle des prénoms, charger les données des départements
1. Donner, pour chaque nom de département, le prénom le plus fréquent depuis l'année 2000

//TODO
val nom = lignes.map(l => (extractField(l, 3),extractField(l, 1),extractField(l, 2).toInt))
val depart = sc.textFile("dpts.txt")
val departement = depart.map(t => (extractFieldS(t, 0),extractFieldS(t, 6)))
val join = departement.join(nom, on="dpt")
//departement.take(10).foreach(println)

//TODO
//val nom = prenoms.filter(l=>l._3>=2000).map(l=>(l._4,(l._2,l._5)))
//val dp = dpts.map(l=>(l._2,l._5))
//val join = nom.join(nom)
//val nbre = join.map(l=>(l._2._1,(l._2._2._1,l._2._2._2)))
//val result = nbre.aggregateByKey(("",Int.MinValue)) ({case((m,n),(v1,v2))=>(if(n>v2) m else v1 , n max v2)},
//          {case((m,n),(v1,v2))=>(if(n>v2) m else v1 , n max v2)})
//result.take(10).foreach(println)
var maxValue = (k:(String,Int), v:(String,Int)) => if ((k._2 max v._2)==k._2 )k else v
var freqByDept= (prenoms.
filter (p=> ((p._3)>1999)).
map(p => ((p._2, p._4),p._5)).
reduceByKey(_+_).
map(p=>(p._1._2, (p._1._1,p._2))).
aggregateByKey(("NONE",Int.MinValue))(maxValue,maxValue)
)
val depart = sc.textFile("dpts.txt")
val result= depart.join(freqByDept).map(d =>(d._2._1._5, (d._2._2._1, d._2._2._2)))
result.take(10).foreach(println)
