==> notebooks/20_formats_de_stockage.ipynb <==
# Des classes pour représenter les données

## Pour les prénoms

case class Prenom(sexe: String, prenom: String, annee: Int, codeDept: Int, nombre: Int)

val ariane = Prenom("2", "ARIANE", 2007, 91, 10)
val cassiopee = Prenom("2", "CASSIOPEE", 2009, 91, 3)
val helios = Prenom("1", "HELIOS", 2012, 91, 1)
println(s"$ariane, $cassiopee, $helios")

## Pour les départements

case class Dept(region: Int, codeDept: String, chefLieu: String, typeNom: Int, nom: String, nomEnrichi: String)

val puyDeDome = Dept(84, "63", "63113", 2, "PUY-DE-DOME", "Puy-de-Dôme")
println(puyDeDome)

# Chargement des données

// Pour les conversions implicites de RDDs vers DataFrames
val sparkRO = spark // bricolage pour que cela fonctionne dans le notebool (inutile sinon)
import sparkRO.implicits._

## Chargement des prénoms

val lignes = spark.sparkContext.textFile("prenoms_sample.txt")
val prenomsRDD = lignes.map(_.split('\t')).map(a => Prenom(a(0), a(1), a(2).toInt, a(3).toInt, a(4).toDouble.toInt))
val prenoms = prenomsRDD.toDS()
prenoms.show() // échoue ici mais fonctionne dans spark-shell ?!?

## Chargement des départements

val lignesDept = spark.sparkContext.textFile("dpts.txt").filter(l => l.startsWith("REGION") == false)
val deptsRDD = lignesDept.map(_.split('\t')).map(a => Dept(a(0).toInt, a(1), a(2), a(3).toInt, a(4), a(5)))
val depts = deptsRDD.toDS()
depts.show() // échoue ici mais fonctionne dans spark-shell ?!?

# Sauve les données au format parquet

## Les prénoms partitionnés par départements et années et compressés (Snappy)

prenoms.write.partitionBy("codeDept", "annee").for
mat("parquet").save("prenomsParDeptsEtAnnees.parquet")

## Les prénoms partitionnés par départements et années et compressés (gzip)

prenoms.write.partitionBy("codeDept", "annee").option("compression", "gzip").format("parquet").save("prenomsParDeptsEtAnnees.gzip.parquet")

## Les départements

depts.write.format("parquet").save("depts.parquet")
