==> notebooks/25_comparaison_des_formats.ipynb <==
# Chargement des données

case class Prenom(sexe: String, prenom: String, annee: Int, codeDept: Int, nombre: Int)

val prenomsRDD = sc.textFile("prenoms.txt").filter(l => l.startsWith("sexe") == false).filter(l => l.contains("XX") == false)

// Pour les conversions implicites de RDDs vers DataFrames
val sparkRO = spark // bricolage pour que cela fonctionne dans le notebool (inutile sinon)
import sparkRO.implicits._

val prenoms = prenomsRDD.map(_.split('\t')).map(a => Prenom(a(0), a(1), a(2).toInt, a(3).toInt, a(4).toDouble.toInt)).toDS()
prenoms.cache

# Sauvegarde dans les différents formats

import java.nio.file._
import java.nio.file.attribute.BasicFileAttributes
import java.util.concurrent.atomic.AtomicLong

def pathSize(path: Path): Long = {
    var size = new AtomicLong(0)

    Files.walkFileTree(path, new SimpleFileVisitor[Path]() {
        override def visitFile(file: Path, attrs: BasicFileAttributes): FileVisitResult = {
            size.addAndGet(attrs.size())
            FileVisitResult.CONTINUE
        }
    });
    size.get()
}

val formats = Map(
    "csv" -> List("uncompressed", "bzip2", "deflate", "gzip"),
    "json" -> List("uncompressed", "bzip2", "deflate", "gzip"),
    "parquet" -> List("uncompressed", "gzip", "snappy"),
    "orc" -> List("uncompressed", "snappy", "zlib")
)

for (format <- formats.keys) {
    for (codec <- formats(format)) {
        prenoms.write.mode("overwrite").option("compression", codec).format(format).save("prenoms")
        val prenomsPath = Paths.get("prenoms")
        println(s"$format, $codec, " + pathSize(prenomsPath))
    }
}

# Affichage des résultats

val resultats = """csv, uncompressed, 68804952                                                     
csv, bzip2, 10676889                                                            
csv, deflate, 11998839          
csv, gzip, 11998875                                                             
json, uncompressed, 238735998                                                   
json, bzip2, 9491007                                                            
json, deflate, 14662059                                                         
json, gzip, 14662095                                                            
parquet, uncompressed, 8344008                                                  
parquet, gzip, 4820150                                                          
parquet, snappy, 6476035                                 
orc, uncompressed, 13265926                                                     
orc, snappy, 5810818                                                            
orc, zlib, 4377612"""

val sizes = resultats.split("\n")

var data = scala.collection.mutable.Map[Tuple2[String, String], Long]()
for (line <- sizes) {
    val cols = line.split(",").map(_.trim)
    data += ((cols(0), cols(1)) -> cols(2).toLong)
}
println(data)

## Taille (en Mo)

val fileFormats = List("csv", "json", "parquet", "orc")
val codecs = List("uncompressed", "bzip2", "deflate", "gzip", "snappy", "zlib")
println("Format | uncomp.  | bzip2    | def.     | gzip     | snappy   | zlib     |")
for (format <- fileFormats) {
    print(f"$format%-7s|")
    for (codec <- codecs) {
        val size = data.getOrElse[Long]((format, codec), 0) / (1024.0 * 1024)
        print(f"$size%9.2f |")
    }
    println
}
