==> notebooks/30_spark_sql.ipynb <==
# Des classes pour représenter les données

case class Prenom(sexe: String, prenom: String, annee: Int, codeDept: Int, nombre: Int)
case class Dept(region: Int, codeDept: String, chefLieu: String, typeNom: Int, nom: String, nomEnrichi: String)

# Chargement des données
1. Créer un dataset pour les prénoms

// Pour les conversions implicites de RDDs vers DataFrames
val sparkRO = spark // bricolage pour que cela fonctionne dans le notebool (inutile sinon)
import sparkRO.implicits._

val prenoms = spark.read.load("prenomsParDeptsEtAnnees.parquet").as[Prenom]
prenoms.createOrReplaceTempView("prenoms")
prenoms.show

1. Faire de même pour les départements

//TODO
val deps = spark.read.load("depts.parquet").as[Dept]
deps.createOrReplaceTempView("deps")
deps.show

# Interroger les données

* La documentation sur Spark SQL est disponible [ici](https://spark.apache.org/docs/latest/sql-programming-guide.html) (de même pour [Dataset](https://spark.apache.org/docs/latest/api/scala/index.html#org.apache.spark.sql.Dataset))
* Pour chaque question, vous donnerez la réponse avec le DSL et en SQL

1. Donner, pour chaque prénom, son nombre d'occurences

## DSL

//TODO
prenoms.groupBy("prenom").agg(sum("nombre")).show

## SQL

//TODO
spark.sql("select prenom, sum(nombre) from prenoms group by prenom").show

1. Donner le nombre total de naissances avec un prénom féminin

## DSL

//TODO
prenoms.where($"sexe">1).agg(sum("nombre")).show

## SQL

//TODO
spark.sql("select sum(nombre) from prenoms where sexe =2").show

1. Donner l'effectif maximal et minimal par prénom

## DSL

//TODO
prenoms.groupBy("prenom").agg(min("nombre"), max("nombre")).show

## SQL

//TODO
spark.sql("select prenom, Min(nombre), Max(nombre)from prenoms group by prenom").show

1. Donner, pour chaque nom de département, le nombre moyen de prénoms depuis l'année 2000

## DSL

//TODO
prenoms.join(deps,prenoms("codeDept")===deps("codeDept")).where($"annee">2000).groupBy(deps("nom"))
.agg("nombre"->"SUM").show

## SQL

//TODO
spark.sql("select avg(p.nombre), first(d.codeDept), d.nom from prenoms p, Deps d where p.annee >= 2000 and d.codeDept=p.codeDept group by d.nom").show
